fix(kernel): the kernel crate type-checks for a VMM-less target
Why: `just check-kernel-portable` (`cargo check -p carrick-kernel --features
test-support --target aarch64-unknown-linux-gnu`) is what proves the extracted
kernel crate compiles on a host with no Hypervisor.framework at all -- the
Linux/KVM, FreeBSD/bhyve and NetBSD/NVMM lanes. It was red on four distinct
type-check defects, every one of them code that had never once been compiled
for a non-macOS target, so nothing about them was load-bearing on macOS.

What:

- `FsView::check_exec_source` / `check_exec_target` (and the private
  `exec_access_errno` that only `check_exec_target` calls) were gated
  `#[cfg(all(target_os = "macos", target_arch = "aarch64"))]` while their
  `SyscallDispatcher` forwarders in `dispatch/fs.rs` were ungated, so off
  macOS the forwarders resolved to nothing (E0599). The bodies are pure
  kernel/VFS logic -- overlay `real_stat`, `canonicalize_following`,
  `layered_lstat`, `resolve_at_path`, `dac_check` -- and name no Darwin-only
  item, so the gate recorded where the code was first needed, not a
  portability fact. Removed, with the stale "Called from `runtime/exec.rs`
  (the macOS/HVF execve path)" comment corrected to the carrier's execve
  path. macOS compiles and behaves exactly as before; the non-macOS lanes now
  get the same `execve(2)` pre-read permission check (execve03 / execveat02 /
  execve02) rather than not having one.
- `OpenDescription::reexec_kind_name` was `#[cfg(any(test,
  all(target_os = "macos", target_arch = "aarch64")))]` beneath an ungated
  caller: the `CARRICK_FICLONE_DEBUG` debug-string arm of `ficlone_errno`. Its
  body is a match over the description variants and names no Darwin-only item,
  so it is now unconditional (and no longer needs `#[allow(dead_code)]`). A
  debug label must not decide portability.
- `flock.l_type` is `c_short` on both hosts, but libc's `F_WRLCK`/`F_UNLCK`
  are `c_int` on Linux and `c_short` on Darwin. `dispatch/ioring.rs`'s
  cross-process io_uring lock now converts at the libc boundary
  (`libc::F_WRLCK as libc::c_short`), exactly as the adjacent
  `l_whence = libc::SEEK_SET as i16` already did. No behaviour moves: the
  value written into the struct field is unchanged on either host.
- `tee_host_passthrough`, the `#[cfg(target_os = "linux")]` arm that had never
  been compiled at all, fed `libc::tee`'s `ssize_t` straight into
  `DispatchOutcome::returned_len_or_errno(len: usize)`. `HostSyscallResult` is
  implemented for `i32`/`isize`/`i64`, so the negative-return errno check now
  happens on the `isize` and the transferred count becomes a `usize` only
  after it -- the same shape the `pwrite` path a few hundred lines down
  already uses.

Verified: `just check-kernel-portable` exits 0 for the first time (7 warnings,
all pre-existing dead-code in code the macOS lane does use, no errors). On the
macOS lane `just fmt-check`, `just clippy`, `just check --workspace`,
`just doc`, `just check-layering` and `just test` (48 suites, 4844 passed, 0
failed, 5 ignored) are all exit 0 -- unchanged counts from before the fix,
which is the claim being made: this moves no macOS behaviour.

+++ crates/carrick-kernel/src/dispatch/ioring.rs
@@ -477,7 +477,7 @@ impl IoUringBacking {
 
     fn cross_process_lock(&self) -> Option<CrossProcessLock<'_>> {
         let mut lock: libc::flock = unsafe { core::mem::zeroed() };
-        lock.l_type = libc::F_WRLCK;
+        lock.l_type = libc::F_WRLCK as libc::c_short;
         lock.l_whence = libc::SEEK_SET as i16;
         lock.l_start = 0;
         lock.l_len = 1;
@@ -518,7 +518,7 @@ impl<'a> CrossProcessLock<'a> {
         }
         self.unlocked = true;
         let mut lock: libc::flock = unsafe { core::mem::zeroed() };
-        lock.l_type = libc::F_UNLCK;
+        lock.l_type = libc::F_UNLCK as libc::c_short;
         lock.l_whence = libc::SEEK_SET as i16;
         lock.l_start = 0;
         lock.l_len = 1;
